requirements.txt
streamlit
neo4j
google-generativeai
pandas
langchain
sentence-transformers
faiss-cpu
